브랜드 추가 기능 완료

+++ admin/src/main/resources/db.h2/schema.sql
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS brand (
 CREATE TABLE product (
      id BIGINT AUTO_INCREMENT PRIMARY KEY,
      name VARCHAR(100) NOT NULL,
-     price INT NOT NULL,
+     price BIGINT NOT NULL,
      category_id BIGINT NOT NULL,
      brand_id BIGINT NOT NULL,
      is_active BOOLEAN DEFAULT true,
